pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "code-repo-learner"
version = "0.2.0"
description = "A tool for learning and analyzing C codebases"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Code Repo Learner Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "tree-sitter>=0.20.0",
    "tree-sitter-c>=0.20.0",
    "pyyaml>=6.0",
    "neo4j>=5.25.0",
    "chromadb>=1.0.13",
    "sentence-transformers>=3.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "colorama>=0.4.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
]

[project.scripts]
code-repo-learn = "code_learner.cli.cli:main"
call-graph = "code_learner.cli.call_graph_cli:main"
dependency-graph = "code_learner.cli.dependency_cli:main"
code-learner = "code_learner.cli.code_analyzer_cli:main"
code-embed = "code_learner.commands.embed_code:main"
code-learner-all = "code_learner.cli.unified_cli:main"

[project.urls]
Homepage = "https://github.com/your-username/code-repo-learner"
Issues = "https://github.com/your-username/code-repo-learner/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"] 

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "mypy>=1.16.1",
    "pytest>=8.4.1",
]
